// src/@types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful';

import { Document } from '@contentful/rich-text-types';

export interface IBlogPostFields {
  /** Title */
  title: string;

  /** Slug */
  slug: string;

  /** Hero Image */
  heroImage?: Asset | undefined;

  /** Description */
  description?: Document | undefined;

  /** Body */
  body?: Document | undefined;

  /** Recommendations */
  recommendations?: Entry<{ [fieldId: string]: unknown; }>[] | undefined;

  /** Professional Experiences */
  professionalExperiences?: Entry<{ [fieldId: string]: unknown; }>[] | undefined;

  /** Volunteer Experiences */
  volunteerExperiences?: Entry<{ [fieldId: string]: unknown; }>[] | undefined;
}

/** Page Data */

export interface IBlogPost extends Entry<IBlogPostFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "blogPost";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IProfessionalExperienceFields {
  /** Organization Name */
  organizationName?: string | undefined;

  /** Role */
  role?: string | undefined;

  /** Location */
  location?: string | undefined;

  /** Dates */
  dates?: string | undefined;

  /** Description */
  description?: string | undefined;
}

export interface IProfessionalExperience
  extends Entry<IProfessionalExperienceFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "professionalExperience";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IRecommendationFields {
  /** Recommendor */
  recommendor?: string | undefined;

  /** Subheadline */
  subheadline?: string | undefined;

  /** Body */
  body?: string | undefined;
}

export interface IRecommendation extends Entry<IRecommendationFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "recommendation";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE =
  | "blogPost"
  | "professionalExperience"
  | "recommendation";

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";